+++ app/src/main/kotlin/com/luminaapps/taigamobile/ui/components/loaders/DotsLoader.kt
@@ -40,8 +40,8 @@ fun DotsLoader() {
         animationSpec = infiniteRepeatable(
             animation = keyframes {
                 durationMillis = delayUnit * 4
-                0f at delay with LinearEasing
-                1f at delay + delayUnit with LinearEasing
+                0f at delay using LinearEasing
+                1f at delay + delayUnit using LinearEasing
                 0f at delay + delayUnit * 2
             }
         )